# openapi.yaml
openapi: 3.1.0
info:
  title: Smart City MVP Ticket System
  version: 1.0.0
paths:
...
  /api/parent-categories:
    get:
      tags:
      - triage
      summary: List Parents
      operationId: list_parents_api_parent_categories_get
      parameters:
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Active Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParentCategoryOut'
                title: Response List Parents Api Parent Categories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - triage
      summary: Create Parent
      operationId: create_parent_api_parent_categories_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentCategoryIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentCategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/categories:
    get:
      tags:
      - triage
      summary: List Categories
      operationId: list_categories_api_categories_get
      parameters:
      - name: parentId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentid
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Active Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryOut'
                title: Response List Categories Api Categories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - triage
      summary: Create Category
      operationId: create_category_api_categories_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/categories/{category_id}:
    patch:
      tags:
      - triage
      summary: Update Category
      operationId: update_category_api_categories__category_id__patch
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: string
          title: Category Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/questions:
    post:
      tags:
      - triage
      summary: Add Question
      operationId: add_question_api_questions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/questions/{question_id}/options:
    post:
      tags:
      - triage
      summary: Add Option
      operationId: add_option_api_questions__question_id__options_post
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          title: Question Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/rules:
    post:
      tags:
      - triage
      summary: Upsert Rule
      operationId: upsert_rule_api_rules_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/categories/{category_id}/questionnaire:
    get:
      tags:
      - triage
      summary: Questionnaire
      operationId: questionnaire_api_categories__category_id__questionnaire_get
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: string
          title: Category Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionnaireOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CategoryIn:
      properties:
        parentCategoryId:
          type: string
          title: Parentcategoryid
        name:
          type: string
          title: Name
        order:
          type: integer
          title: Order
          default: 0
        active:
          type: boolean
          title: Active
          default: true
        defaultRecommendationText:
          type: string
          title: Defaultrecommendationtext
          default: ''
        defaultUrgency:
          $ref: '#/components/schemas/UrgencyLevel'
          default: MEDIUM
        organizationIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Organizationids
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - parentCategoryId
      - name
      title: CategoryIn
    CategoryOut:
      properties:
        id:
          type: string
          title: Id
        parentCategoryId:
          type: string
          title: Parentcategoryid
        name:
          type: string
          title: Name
        active:
          type: boolean
          title: Active
        order:
          type: integer
          title: Order
        defaultRecommendationText:
          type: string
          title: Defaultrecommendationtext
        defaultUrgency:
          $ref: '#/components/schemas/UrgencyLevel'
        organizationIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Organizationids
      type: object
      required:
      - id
      - parentCategoryId
      - name
      - active
      - order
      - defaultRecommendationText
      - defaultUrgency
      title: CategoryOut
...
    CategoryUpdateIn:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
        order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Order
        defaultRecommendationText:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultrecommendationtext
        defaultUrgency:
          anyOf:
          - $ref: '#/components/schemas/UrgencyLevel'
          - type: 'null'
        organizationIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Organizationids
      type: object
      title: CategoryUpdateIn
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OptionIn:
      properties:
        label:
          type: string
          title: Label
        code:
          type: string
          title: Code
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - label
      - code
      title: OptionIn
    OptionOut:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        code:
          type: string
          title: Code
      type: object
      required:
      - id
      - label
      - code
      title: OptionOut
    ParentCategoryIn:
      properties:
        name:
          type: string
          title: Name
        order:
          type: integer
          title: Order
          default: 0
        active:
          type: boolean
          title: Active
          default: true
        defaultUrgencyHint:
          $ref: '#/components/schemas/UrgencyLevel'
          default: MEDIUM
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - name
      title: ParentCategoryIn
    ParentCategoryOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        order:
          type: integer
          title: Order
        active:
          type: boolean
          title: Active
        defaultUrgencyHint:
          $ref: '#/components/schemas/UrgencyLevel'
      type: object
      required:
      - id
      - name
      - order
      - active
      - defaultUrgencyHint
      title: ParentCategoryOut
    QuestionIn:
      properties:
        categoryId:
          type: string
          title: Categoryid
        text:
          type: string
          title: Text
        order:
          type: integer
          title: Order
          default: 0
        active:
          type: boolean
          title: Active
          default: true
        options:
          items:
            $ref: '#/components/schemas/OptionIn'
          type: array
          title: Options
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - categoryId
      - text
      title: QuestionIn
    QuestionOut:
      properties:
        id:
          type: string
          title: Id
        categoryId:
          type: string
          title: Categoryid
        text:
          type: string
          title: Text
        order:
          type: integer
          title: Order
        active:
          type: boolean
          title: Active
        options:
          items:
            $ref: '#/components/schemas/OptionOut'
          type: array
          title: Options
      type: object
      required:
      - id
      - categoryId
      - text
      - order
      - active
      - options
      title: QuestionOut
    QuestionnaireOut:
      properties:
        category:
          $ref: '#/components/schemas/CategoryOut'
        parent:
          anyOf:
          - $ref: '#/components/schemas/ParentCategoryOut'
          - type: 'null'
        questions:
          items:
            $ref: '#/components/schemas/QuestionOut'
          type: array
          title: Questions
        rules:
          items:
            $ref: '#/components/schemas/RuleOut'
          type: array
          title: Rules
      type: object
      required:
      - category
      - questions
      - rules
      title: QuestionnaireOut
    RuleIn:
      properties:
        categoryId:
          type: string
          title: Categoryid
        match:
          additionalProperties:
            type: string
          type: object
          title: Match
        recommendationText:
          type: string
          title: Recommendationtext
        priority:
          type: integer
          title: Priority
        active:
          type: boolean
          title: Active
          default: true
        setsUrgency:
          anyOf:
          - $ref: '#/components/schemas/UrgencyLevel'
          - type: 'null'
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - categoryId
      - match
      - recommendationText
      - priority
      title: RuleIn
    RuleOut:
      properties:
        id:
          type: string
          title: Id
        categoryId:
          type: string
          title: Categoryid
        match:
          additionalProperties:
            type: string
          type: object
          title: Match
        recommendationText:
          type: string
          title: Recommendationtext
        priority:
          type: integer
          title: Priority
        active:
          type: boolean
          title: Active
        setsUrgency:
          anyOf:
          - $ref: '#/components/schemas/UrgencyLevel'
          - type: 'null'
      type: object
      required:
      - id
      - categoryId
      - match
      - recommendationText
      - priority
      - active
      title: RuleOut
...
    UrgencyLevel:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      title: UrgencyLevel
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer